// web/src/api/generated/models/RelatedPersonDeletePolicy.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const RelatedPersonDeletePolicy = {
    preserve: 'preserve',
    cascade: 'cascade'
} as const;
export type RelatedPersonDeletePolicy = typeof RelatedPersonDeletePolicy[keyof typeof RelatedPersonDeletePolicy];


export function instanceOfRelatedPersonDeletePolicy(value: any): boolean {
    for (const key in RelatedPersonDeletePolicy) {
        if (Object.prototype.hasOwnProperty.call(RelatedPersonDeletePolicy, key)) {
            if (RelatedPersonDeletePolicy[key as keyof typeof RelatedPersonDeletePolicy] === value) {
                return true;
            }
        }
    }
    return false;
}

export function RelatedPersonDeletePolicyFromJSON(json: any): RelatedPersonDeletePolicy {
    return RelatedPersonDeletePolicyFromJSONTyped(json, false);
}

export function RelatedPersonDeletePolicyFromJSONTyped(json: any, ignoreDiscriminator: boolean): RelatedPersonDeletePolicy {
    return json as RelatedPersonDeletePolicy;
}

export function RelatedPersonDeletePolicyToJSON(value?: RelatedPersonDeletePolicy | null): any {
    return value as any;
}

export function RelatedPersonDeletePolicyToJSONTyped(value: any, ignoreDiscriminator: boolean): RelatedPersonDeletePolicy {
    return value as RelatedPersonDeletePolicy;
}